Trying to answer power of as double

+++ src/main/java/com/oocode/Answerer.java
@@ -18,7 +18,7 @@ public class Answerer {
             } else if (Objects.equals(questionList.get(3), "minus")) {
                 return String.valueOf(Integer.parseInt(questionList.get(2)) - Integer.parseInt(questionList.get(4)));
             } else if (Objects.equals(questionList.get(5), "power")) {
-                return String.valueOf((int) Math.pow(Double.parseDouble(questionList.get(2)), Integer.parseInt(questionList.get(7))));
+                return String.valueOf(Math.pow(Double.parseDouble(questionList.get(2)), Integer.parseInt(questionList.get(7))));
             }
         } else if (question.contains("Which of the following numbers is the largest:")) {
             List<String> questionList2 = Arrays.asList(question.split(":"));
